pyproject.toml
[project]
name = "academic-system-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.3.1",
    "dotenv>=0.9.9",
    "ipykernel>=7.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-tavily",
    "langgraph>=1.0.3",
    "pymupdf>=1.26.7",
    "typing>=3.10.0.0",
    "requests",
    "feedparser",
    "fastmcp",
    "mcp[cli]>=1.6.0",
    "PyPDF2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0", # Better HTML parser for BeautifulSoup
    "httpx[socks]>=0.28.1",
    "langchain>=1.1.3",
    "langchain-mcp-adapters>=0.2.1",
    "loguru>=0.7.3",
    "pdfplumber>=0.11.8",
    "fastapi>=0.127.0",
    "sqlalchemy>=2.0.45",
    "faiss-cpu>=1.13.1",
    "langchain-openai>=1.1.3",
    "langchain-milvus>=0.3.2",
]
